minor ui improvement
